requirements.txt
sanic
requests
python-dotenv
sentry-sdk